Defined DeviceTypeSubclass type
For cases where we want to associate osd and pools with devices on
specific nodes with specific devices, we need a more fine-grained level
of device-type, which is passed on to rook as CRUSH-device-class.

As an example, consider cluster with ssd devices from two different
vendors, where user would want to have different pools, one for each
sub-device type. When combined with node-labeling and this patch, the
task becomes trivial.

Signed-off-by: Shachar Sharon <ssharon@redhat.com>